+++ src/__tests__/views/TransactionsPage/TransactionsPage.test.js
@@ -1,17 +1,9 @@
 import { fireEvent, screen, waitFor } from "@testing-library/react";
 import { renderWithProviders } from "../../__test-utils__/test-utils";
 import { rest } from "msw";
-// eslint-disable-next-line jest/no-mocks-import
 import { server } from "../../__mocks__/api/server";
 import GenericPage from "../../../views/GenericPage/GenericPage";
 import { ENTITY_TRANSACTIONS } from "../../../utils/constants/constant";
-
-beforeEach(() => { server.listen(); });
-// Reset any request handlers that we may add during the tests,
-// so they don't affect other tests.
-afterEach(() => { server.resetHandlers(); });
-// Clean up after the tests are finished.
-afterAll(() => server.close());
 
 describe("Test TransactionsPage", () => {